requirements.txt
amqp==5.1.1
asgiref==3.7.1
async-timeout==4.0.2
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-admin-inline-paginator==0.4.0
django-admin-interface==0.26.0
django-admin-rangefilter==0.10.0
django-colorfield==0.8.0
django-constance==2.9.1
django-environ==0.10.0
django-filter==23.2
django-image-uploader-widget==0.2.5
django-import-export==3.2.0
django-json-widget==1.1.1
django-phonenumber-field==7.1.0
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dlib==19.24.2
elastic-transport==8.4.0
elasticsearch==8.8.0
emoji==2.5.0
et-xmlfile==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fake-useragent==1.1.3
future==0.18.3
idna==3.4
joblib==1.2.0
kombu==5.2.4
Markdown==3.4.3
MarkupPy==1.14
memory-profiler==0.61.0
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.24.3
odfpy==1.4.1
opencv-python==4.7.0.72
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pathspec==0.11.1
pdfkit==1.0.0
pdfminer==20191125
phonenumbers==8.13.11
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.6
pycryptodome==3.18.0
Pygments==2.15.1
PyJWT==2.7.0
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
redis==4.5.5
regex==2023.6.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tablib==3.4.0
text-unidecode==1.3
tqdm==4.65.0
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
wkhtmltopdf==0.2
xlrd==2.0.1
xlwt==1.3.0
